Document Jinja2 templating.
