+++ betty/extension/cotton_candy/assets/templates/entity/featured--place.html.j2
@@ -17,7 +17,7 @@
 </div>
 
 <div class="featured-entity-feature">
-    {% set places = place | walk('encloses') | select('place_entity') | list %}
+    {% set places = place | walk('encloses') | select('entity', 'Place') | list %}
     {% if place.coordinates %}
         {% set places = places + [place] %}
     {% endif %}
