+++ betty/extension/cotton_candy/assets/templates/entity/page--place.html.j2
@@ -4,7 +4,7 @@
 {% block page_content %}
     {% include 'entity/meta--place.html.j2' %}
 
-    {% set places = place | walk('encloses') | select('place_entity') | list %}
+    {% set places = place | walk('encloses') | select('entity', 'Place') | list %}
     {% if place.coordinates %}
         {% set places = places + [place] %}
     {% endif %}
